+++ match3_elements_unity/Assets/Scripts/Meta/MainWidget/MetaMainWidgetGameRegimeSyncStartAction.cs
@@ -6,7 +6,7 @@ namespace Meta.MainWidget
 {
 	public sealed class MetaMainWidgetGameRegimeSyncStartAction : WidgetGameRegimeSyncStartAction<IMetaWidgetProvider>
 	{
-		public override int priority => (int)MetaGameRegimeSyncStartActionPriority.MainWidget;
+		public override byte priority => (byte)MetaGameRegimeSyncStartActionPriority.MainWidget;
 
 		[Inject]
 		public MetaMainWidgetGameRegimeSyncStartAction(IMetaWidgetProvider widgetProvider) : base(widgetProvider) {}
